Rename labels_stringlabels.go -> labels.go
Since labels_stringlabels.go is now the default implementation we should rename label files to make this more obvious.
